[850] Add incremental fetching to course controller
This allows users to pass in a changed_since parameter in order to
specify a range of courses to fetch.

Also refactors the next_link_header logic into its own concern, as
it's shared with the provider controller.

Use the from_course_id parameter to get the correct set of results,
which allows us to paginate when multiple records have the same timestamp.

A concern is also created for the FirstItemFromNextPage logic, that
will be pulled into the providers controller in the near future
(Trello #968).