[995] Re-implement find, apply and vacancy checks on Course
This commit moves the logic for checking if a course is on Find,
whether it's open in Apply and whether it has vacancies into ActiveRecord,
using chaining of the enum scopes.

As the functionality is already tested in the Course spec, it's redundant
to check the scopes exhaustively again in the SiteStatus spec, so these
are removed.