[875] handle PG connection errors in ActionDispatcher
This was originally implemented in the controller, but we realised that this
will be hiding stack traces in development and is better implemented in
ActionDispatch (which is the way Sau oroginally wanted to do it1). We've
reverted to this implemention and added a super-simple (naive) configuration
spec to test it.