Use auth token from environment if available in dev
UCAS env is now in dev mode, so it had reverted to the default token
which is not what we want.

This patch means that local development still gets the default token,
and we can run development mode in azure with a custom token (for
pre-production environments).

This will go away when we switch to the settings gem for the v1 api.