[1406] Improve course enrichment validation messages
Add `custom_error_message` as a dependency. It allows us to prefix error
messages with `^` when they should not use the attribute as a prefix
for the full_message.

This modifies the `course#validate_enrichment` method. Before, all
enrichment errors would be hoisted under the `latest_enrichment` key.
This isn't enough information to be able to determine which field this
error belongs to. The new key is computed as `latest_enrichment__FIELD`.

Updated tests and adds better validation messages for course
enrichments via `en.yml` overrides.