[1413] Skip email update if email is already taken
The database constraint for email uniqeness on users causes ActiveRecord
to raise an error whenever we try to update the email to one that
already exists.

This happens for some users that have a DfE-SignIn ID on a user record
with an email address different from the one that we get from
DfE-SignIn, and the email from DfE-SignIn matches another record.