[1413] Send exception to Sentry for duplicate user
DfE-SignIn gives us a payload with an ID and an email.

When we look up the user by the ID and the email doesn't match the
user's email, we want to update their email to the one we get from
DfE-SignIn.

Sometimes though, there's already another user on the system with the
same email, so we get an error. Instead of just not updating and moving
on with authenticatation, we want to send an exception to Sentry with
the details of the users.

This allows us to start collecting some data about the users affected by
this, which will help us come up with a plan to resolve this, wheter
that is manually or automatically.