[1472] Delete site status when removing site for new courses
If a publisher adds and then removes a training location on a course
before course is published, they wouldn't expect it to come back once
they've set the course to running.

To achieve this, we need to delete the site status, but only when the
course is new. If it's being removed post-publication, it should be suspended
as before.