[1515] Change course and subject relationships
Has and belongs to many introduces as a lot of magic to make it work.
After we introduced bullet, it wasn't working correctly with this
because of the custom join table.

To fix this, we've made CoursesSubject a model and configured the table
name to use the table we have in the database.

This allows us to replace the HABTM with a has many through
relationship, which works better in this case.