[1527] Training locations without vacancies should be suspended
If a training location has no vacancies, it's still selectable on UCAS Apply
but the candidate is informed several steps into their application journey
that it's full.

This change makes the training location suspended in the V1 API. The end effect
of this change is that the location is no longer listed on UCAS Apply, so candidates
don't inadvertently select them anymore.