[1580] Add validations to edit locations
This will allow us to display an appropriate error message on the
frontend when the user submits an empty array of sites.

The validation is done at the controller level instead of the
model. This is because `sites = []` is something that we can
validate against, but we can't actually revert easily from what I can
tell because of the `remove_site!` side effects that occur when it's
called.