[1587] Create enrichment in courses#update only when necessary
Currently the courses#update action always creates an enrichment when
called. This leads to two bugs:

- When updating the sites, the course is set to draft
- When the user opens the enrichment pages but makes no changes and
  hits "Save", the course changes state

This wraps the enrichment logic in a guard so that it isn't run
unnecessarily.

The enrichment and site update logic are both wrapped in private
methods to make the update action a bit easier to parse.