[1455] Use JSONAPI matchers instead of equality
This test was too specific, checking each and every attribute. This
caused it to be too brittle.

I've replaced comparing the complete JSON output to using JSONAPI RSpec
matchers to check if the correct objects are rendered out, with the
relationships and attributes we exepect.

There's no need to test this further than that, the serializer is unit
tested to make sure it gives us the correct data.