Revert "Merge pull request #451 from DFE-Digital/1598-use-callbacks-to-add-or-remove-course-sites"
This reverts commit cc1eadf5c099861f9ac2efe272fd6e742c2fc6bc, reversing
changes made to 07c22b2d36931e6a18fb56ba557c4e2ddc186aae.

A nasty, gnarly bug was discovered which hadn't been tested for. When using
callbacks and the built-in has-many-through mechanism it seems to be impossible,
in a practical sense, to modify existing sites statuses back to "running" when
being created instead of instantiating new ones. As we are likely to be
modifying the data model to remove the use of these statuses anyway, reverting
this and focusing on revising the data model seems like the best use of time.