[1624] Remove latest_enrichment prefix on enrichment errors
These were initially added with the idea that without them, the frontend
will not be able to differentiate which errors belong to the course and
which errors belong to the course enrichment.

In practice, they actually end up confusing json-api, which will not
automatically assign `source: { pointer }`s which are used to then
associate validation errors to the appropriate field.

Because the frontend doesn't need to know about enrichments, it's better
if we simply remove this prefix.