[1664] Override reverse_mapping in deserializable_course
`reverse_mapping` is responsible for generating the `pointer`
attributes in a JSON API response which contains validation
errors. The way it works is `jsonapi-rails` will inspect the
`params` and based on the attributes provided, populate
`reverse_mapping`. This means that the only attributes that can
be associated with errors are ones that get passed in as part of
the request.

When we do the `publish` action, we don't pass in any attributes,
because they already exist on the model. This means that without
help, `jsonapi-rails` can't figure out to which attributes the fields
we're showing errors on should map back to. In this scenario, it
silently moves on without setting pointers.

We need validation errors to always have pointers, so it makes sense to
force it to always include them by overriding the default
`reverse_mapping`.

The qualifications are also added to the `attributes` call, to allow
them to correctly map.