Expose edit options for entry requirements
Backend should be responsible for the business logic which determines which options are available for each course. Eg which course outcomes should be listed when editing the outcome.

We need a place for these and we need to share them with the frontend.

Add an EditOptions class and use it pass through the options for editing entry requirements.