[1997] add endoint for new course
From the ticket:

Add "new" endpoint for courses to mcbe that returns default `edit_options` to be
used when generating the "new" pages on the frontend.

We are keeping the logic behind course states constrained to the backend, so we
need to be able to query it to find out what outcomes are available, for courses
as we step the user through the wizard steps.