[2055] Add course_builder endpoint to v2 api
This is not a JSON::API endpoint because:

I've read the JSON API spec https://jsonapi.org/format/ and it appears
to conflict with the requirements of our build_new endpoint. Notably:

*  Requires id on resources (objects) returned from the server (we
aren't persisting anything so there is no id).
*  Only supports the four CRUD operations, to the exclusion of all
else, so even though meta is available there is no request we could
attach our generated course to.

I've talked this over briefly with @misaka and we've agreed to ditch
json-api for this and create a vanilla json endpoint specifically for
course building.