[2258] Only sync published course on course update
If a course hasn't been published yet, it's not on Find, and we don't need to sync it.

This wasn't true last cycle, in 2019 a course was on Find if it had been set to running on UCAS – a course's published content was published separately.

From 2020 the only way a course gets set to running is by publishing it. So we can now include this check.

This will prevent errors when trying to sync something in a non-syncable state (eg no findable sites)