[2313] memoize Subject.modern_languages
This is an optimisation for the bulk sync courses load. It also requires changes
to tests to ensure the modern languages object used throughout is consistent,
which means creating a modern language object in the db outside of the spec
transactions. The simplest way to do this is to use the SubjectCreatorService
which means all the subjects get created outside spec transactions. This is an
optimisation that could be useful to use for subjects and recruitment cycles in
general, and better to do it now before a whole bunch of tests have to be fixed
to accomodate it.