Change dev psql user to match production db dumps
This will remove a step for onboarding new developers when restoring a
sanitized backup of production data.

The production db uses manage_courses as a username, and that is encoded
in the production databaes dumps, without this patch you get errors
restoring the production db dumps because the manage_coures role does
not exist, meaning you have notice in the scrollback and manually create
the role.

This patch will mean all devs will have to re-run `rake db:setup` and
optionally manually drop the old role.