Add COUNTY_TO_REGION_CODE initializer
This is a copy of a list of UK counties and their respective regions, plus some data munging. The data munging is done in code (as opposed to hard-coded in the initial array) to make the operations that were done more transparent. The data itself will probably not change.

Ignore this file from Rubocop because using %w() is a really bad idea in
this case.