Add Site.with_locatable_address
Scope that only returns sites that have an address that has either a
first line or a postcode. This is a temporary measure to deal with
historic Site data that has invalid addresses. Validation is in place
now to prevent this but while the data is invalid we need to omit sites
from search results.