[3116] Move attribute conversion into the mailer
Formatting is required before values can be added to the email template.
We don't want to couple these to the 'translated' values from the locale
file so for now move the translation into the mailer so as we have the
original attribute name available.

Further refactoring needs to be done here to remove the email
notification logic from the `Course` model altogether.