[3033] Remove .rspec file
This file was requiring spec_helper for all specs. This isn't necessary
because:

a) spec_helper is required explicitly in all of the specs anyway
b) rails_helper is being required in some of the files that spec_helper
requires so this ends up also requiring rails_helper which loads the app
and this isn't what we want
c) this stops the pre_deploy_spec being run on a docker container
without db access which we want to do.

So, remove it.