[3104] refactor api v2 organisation request specs
These tests are failing because the order of the returned results are changing
on this branch. So change the tests not to expect realised JSON response and use
more targetted matchers.

This does mean that we don't have a JSON version of the response in a place
that's handy to access, but if we want that maybe we should document the API
with open API, or failing that we can still spit out the response JSON inside of
the tests. We can address that need some other way.