[3216] Remove vnet config from template
When we deploy, the route table added by CIP is removed. It is required to
use the Internet Access Service (IAS). This is because our Azure template
also configures vnet and subnet, but without route table.

Now we use a simpler network template which does not try to recreate the
network config, but only references the existing.