[3307] Add `request_type` support to API
Initial work to support `request_type` parameter on the Allocation API.

Set the number of places from the request_type. Ulimately only `initial`
request type will be allowed to supply an explicit `number_of_places`
but that will be added in a subsequent PR once Publish is sending the
correct parameters.