Fix flakey organisation spec
Don't test the whole JSON response in the v2 organisation request spec.

This spec fails quite regularly as the ordering is flakey within the
JSON. We have moved to not testing the entire JSON response in request
specs as this should be well covered in the serializer specs so removing
that here should make the tests more reliable.