[3364] Don't use ProviderPolicy in /courses
The `/api/v2/courses` endpoint was checking provider policy if the
`provider_code` was present in the params. This logic should be defined
in the course policy and is already present in the scope so rely in that
instead.

We need to fetch courses for a provider that have a particular accredited
body as a user of that accredited body. The existing implentation did
not apply filters if the provider code was supplied and relied on the
provider policy to enforce authorization which meant the user would get
a 403.