[3393] Add UserNotificationPreferences
Object that represents the preferences chosen (or inferred) relating to
a users notifications.

This is used currently for users to indicate that they wish to be
informed when a course is updated or created by a provider that has a
relationship to an accredited body that the user belongs to.

This class isn't an ActiveRecord model but provides a cleaner interface
to the collection of UserNotification records that actually persist the
data.

`id` is actually the users `id`.

The class also provides an `updated_at` field to allow the client app to
provide different UI if the user has never made an explicit choice.

Co-authored-by: Aga Dufrat <aga.dufrat@digital.education.gov.uk>