[3417] don't validate courses when rolling over
This may seem counter-intuitive, but it seems that our courses aren't all valid
at the current time, for example the "age_range_in_years" is nil on a bunch of
courses, but is validated as being present on new/create. We should fix this,
but for the time being we just need to duplicate the broken-ness.

This change also adds a little extra info when an error is raised to know which
course code was being processed at the time of the error.