[202] Add permission_given field to Contact
We require the user to give explicit permission when saving a contact.
This field allows us to explicitly capture that and return and error
(and prevent creation/update) if permission has not been checked in the
UI.

Permission has to have been given for existing contacts as they have
been manually added as a support request so set those to true.