Fix failing spec relating to rollover
We are currently in a weird state between closing the current_recruitment_cycle
and opening up the next_recruitment cycle. This caused a problem with
one test failing when it tried to save a findable open course. It
was failing because the tests were setting the start date and the
application_from_date as today (using a date outside the
current current_recruitment_cycle) instead of using the date set
by the factories.