[239] Fix rollover reporting information
There was an error in Publish performance dashboard where the page failed
to load because the next recruitment cycle is not created until we do
rollover again.

We decided that this should really be handled the api and instead of
attempting to count data that doesn't exist we decided that the guard
would instead return a default data structure with zero out figures.

We tested this locally and this worked with publish because the rollover
tab is hidden from the user until the next cycle is created and we
are in rollover state.

Co-authored-by: SamJamCul <samuel.culley@digital.education.gov.uk>
Co-authored-by: Jonathan Filar <58793682+JonF1@users.noreply.github.com>