[252] Add includes to providers#show
This endpoint has performance issues.

It is only currently called with one set of JSON API includes so hard
code the DB includes that are required. This should generally be
dynamic if we're supporting json-api fully and set in response to the
includes and fields requested by the caller but our data model makes
that complex.

Hard code 'em.