[252] Add includes for course index
Include relationships in the course query in `v3::CoursesController#index`

The CourseSearchService handles this query. This commit adds includes to
the outer query scope within the service. This reduces the N+1 issues
but doesn't relate to the includes parameters accepted at the
controller. This endpoint is used by Find so we probably don't need that
kind of flexibility.