[506] Pagination params are nested
Our documentation defines pagination params as:

page: { page: <whatevs>, per_page: <whatevs> }

In `ApplicationController` we don't handle this and expect the `page`
and `per_page` to be in the root of the params.

This commit handles both cases so behaves as documented but doesn't break
any clients that have already worked around.