[993] Add sandbox user rake task
Rake task to allow creation of users from a CSV in the sandbox
environment. The csv can be piped into STDIN from a local machine and
into this rake task on a running instance of the API in sandbox.

This task doesn't currently create providers as we need to have a think
of the best way to handle the provider type etc and the implications of
that for sandbox users. This will get us started with the publish
sandbox for now.