Add `published` scope to V3 courses index endpoint
This makes it consistent with the show action in the same controller so
that we don't get into a situation where a course is returned by the
search API (index action) but then 404s for the show action.

Also fixes spec regression - Can't use the same CourseEnrichment object
on two different courses.