[2031] Add debug info to DuplicateUserError
It is impossible to debug this without more information.

Sentry just prints out the error message for the exception, (we could
also use something called raven_context to add some json to the 
exception, but that feature is undocumented so this feels safer). It
would be nice to have the email address as well, but this would be 
filtered out by our sentry config for PII reasons.