Add legacy support for required_qualifications field in V1 public API
We've updated course creation in Publish to take structured data for
required degree and GCSE information. This will eventually replace the
free text `required_qualifications` attribute.

To maintain the existing interface of the V1 public API, we need to
transform the structured data into a value that can be returned in the
`required_qualifications` field in scenarios where a Course has no value
for that attribute.

Do this by adding a RequiredQualificationsSummary object that's invoked
via the Course model. This object:

- Takes an instance of Course as an argument
- Looks for the `required_qualifications` enrichment attribute, and uses
  it if it has a value
- Otherwise generates summary text from the qualifications-specific
  attributes on Course
- Handles just GCSEs for now, with degrees to follow in a later commit