Change how required_qualifications are serialized for V1 public API
In the course serializer for this API, instead of relying solely on the
required_qualifications enrichment attribute, call the
`Course#required_qualifications` method instead.

This in turn invokes `RequiredQualificationsSummary` which retrieves the
appropriate value depending on whether or not the course captures this
information as structured data or free text.