[2187] Previously published courses rollable
* Update `Course#rollable?` to include previously published courses
* Refactor `RolloverService` spec

Courses that have been published and then subsequently changed (without
then publishing) should be rolled over. Change the definition of
`rollable` to include these courses.

This change broke the `RolloverService` spec despite that service not
being concerned with what is `rollable` or not so refactor to test what
the service is actually doing i.e. fetching courses and passing them to
other services that are concerned with `rollable`.