[2392] Throttle on Redis out of memory
* Retry 5 times on Redis::CommandError
* Sleep for 10 seconds to give Sidekiq time to clear some jobs

Redis will raise a `Redis::CommandError` when it reaches its configured
memory limit. When we ran on staging we almost reached that limit so as
an extra bit of safety this commit will cause a sleep to allow Sidekiq
to reduce the queue size back below the threshold. The staging data is
effectively identical to production so it is unlikely that we'll hit
this limit with more workers running in production. This should be
sufficient to deal with an unexpected peak.