Optimisations for /v3/courses SQL - extract nested query
Attempt to remove the sub-query that was introduced to avoid duplicates
when filtering by provider name. The sub-query applies all the filters
and just fetches the `course.id`s. The outer query fetches the data,
handles pagination and sorting. The theory here is that we can combine
the two queries and potentially make a performance gain in doing so.
I've added a `DISTINCT(course.id)` to the query to avoid any duplicate
records.

All needs properly testing to make sure it's returning the correct data.
Also need to properly measure any performance gain that there might be.