Optimisations for /v3/courses SQL - remove double join
Attempt to avoid the double join to `site_status`. These seems to be the most
expensive parts of the query. I've changed the `within` and
`by_distance` scopes to join to `site` via `site_status` so that
when they are combined with other scopes that reference `site_status` they
can share the same SQL join.

All needs properly testing to make sure it's returning the correct data.
Also need to properly measure any performance gain that there might be.