Attempt to fix duplicates when searching by provider name
This is an attempt to fix the `PG::InvalidColumnReference: ERROR: for
SELECT DISTINCT, ORDER BY expressions must appear in select list` error
that happens when searching by provider and attempting to de-dup with a
`SELECT DISTINCT(course.id)`. By bringing the `CASE` expression from the
`Provider.by_provider_name` scope into the `SELECT` list for the main
query.

Tests are broken again.