Optimisation for V3::CourseSearchService SQL - merge inner/outer query
Attempt to remove the sub-query that was introduced to avoid duplicates
when filtering by provider name. The sub-query applies all the filters
and just fetches the `course.id`s. The outer query fetches the data,
handles pagination and sorting. The theory here is that we can combine
the two queries and potentially make a performance gain in doing so.
I've added a `DISTINCT(course.id)` to the query to avoid any duplicate
records.

Based on 3ed55ca689, copied here to account for the original having been
implemented on a different file.

Co-authored-by: Steve Hook <steve.hook@gmail.com>