Change how V3::CourseSearchService deals with provider ordering
The existing change to Course.accredited_body_order, made to optimise
the V3 search service, doesn't play well with other clients of this code
(due to the missing 'by_provider_name' column, which is only created via
a SQL 'AS').

Move the required modification straight into the search service, leaving
the original scope on the Course model as it was before.