Fix Stats service to pass in correct cycle
The statistics for allocations on the performance Dashboard are coming
through incorrectly. After looking into this the service is workingas it
should be. It's just that because Allocations take place in June/July
i.e. before cycle starts they are associated with previous year's
recruitment cycle. The quick fix here is to pass in the previous year.
I've done this and run the numbers past Adam who says they all look as
expect for the previous and current cycles.