Fix flaky RequestStore spec
The addition of the Draper gem to the project in an earlier commit meant
that a 'current_controller' value gets stored in RequestStore on certain
requests. This broke a request spec which asserts the presence of
request_id.

The spec broke because it uses a mock object with strict expectations
about what messages it receives. Just use a hash instead and check its
values after calling the relevant endpoint.