Tweak error message formatting in About Your Organisation
The presence of the 'custom_error_message' gem in this project means that
there's a convention throughout TTAPI of prefixing custom errors with a
carat symbol ('^'). The gem overrides 'full_messages' to allow these
errors to be displayed without the attribute name prefixing them.

There's probably a better way to handle custom errors, but for now we
use full_messages in the AboutYourOrganisationForm so that we aren't
displaying the carat symbol in the UI when rendering errors.