Remove visa-related lines from provider details feature spec
Testing of the visa section will be handled in a different spec so we
can remove references to the functionality from this one.

We add the 'Provider#can_only_sponsor_skilled_worker_visa?' method
because it's needed under certain conditions by a helper that's used
when rendering the details show page. Not adding it means random
failures if we don't explicitly set values for the provider visa fields
(because these are assigned randomly by the factory).