Change header implementation for Publish interface
Match the markup, styling, and content used in Old Publish, so that
users have a consistent experience as they move between Old and New.

This uses the Old Publish implementation as a starting point and moves
across the relevant pieces into the component-based header we have in
this codebase.

This also involves adding a phase banner component, as Old Publish
displays this underneath the main header, rather than as an inline label
in the header itself.