Use Old Publish paths in the header
While we're migrating code from Old Publish to New Publish, keep links
in the header pointing to routes on Old Publish.

When notifications gets migrated across, its link can be updated.
Similarly, one of the last steps of the migration should be to change
the homepage link to point locally.