Add link to Old Publish on organisations page
ProvidersController#index (the 'Organisations' page) is not currently
implemented in New Publish - it's currently an empty page. The plan is
to migrate the various sections across and then do this root page last.

If a New Publish user ever ends up navigating to
'/publish/organisations', they need a way back to Old Publish. This
shouldn't typically happen, but is an edge-case that can arise when the
user explicitly signs out and signs back in on New Publish.

We also don't want to just redirect them to Old Publish in the
controller, as we need some kind of root page in the local app to act as
the default after sign in path. Having one also makes authenticated user
flows easier to write in our feature specs.

Given all this, let's add a simple link to the page which takes them to
the equivalent path on Old Publish.