Changes to backup process
The overnight backup will now be used as part of the DR process. The
timestamp has now been removed to make the DR process easier to execute.

Backups are now created 7 days a week.

Users will be warned that the earlier backup from that day will be
overwritten when trying to manually run the backup process.