ignore cloudfront proxy IPs for request.remote_ip
Up until now, when retrieving the IP of the client we've been getting
the CloudFront proxy IPs. The cloudfront-rails gem retrieves and caches
the list of CloudFront IPs, and configures Rails to ignore them when
retrieving the client IP via request.remote_ip.

This is required by dfe-analytics to help us better understand who is
using our services (using psuedo-anonymisation).