Distinguish between published and non published changes
The published timestamp will not be set on newly created draft
enrichments. The lack of a published timestamp is then used as part
of the logic to determine if changes have been published or not.

Find views now make use of the new published decorator methods
that rely on the aforementioned timestamp.