[549] Fix cycle switcher
This was not working once we were in one of the between periods. The if statement would always return true for the first one because the second part of this or statement would always be true.

`current_cycle_schedule == :today_is_after_find_opens || Time.zone.now.between?(find_opens, apply_2_deadline)`