Set a default value of false for visa sponsorship
We don't set a value in these two fields when we onboard a new
provider, yet we are validating for the presence of these
values in order for the provider to be able to publish. We also
offer no way of the provider being able to set these values
themselves.