Don't rely on fallback behaviour in Rswag specs
The Rswag spec for the API endpoints:
- `/recruitment_cycles/{year}/providers`
- `/recruitment_cycles/{year}/courses`

Were silently dependent on fallback behaviour where the current
recruitment cycle is returned when the specified year doesn't exist.

This was happening because the specs didn't explicitly create a
recruitment cycle to test against.

To test this, I temporarily disabling the fallback behaviour in the
associated controllers:
- Line 89 of `app/controllers/api/public/v1/providers_controller.rb`
- Line 29 of `app/controllers/api/public/v1/courses_controller.rb`

However I haven't committed the controller changes to avoid making any
functional changes and breaking existing behaviour.