Handle requests for providers in the current year
This commit tightens up the handling of requests for providers.

Previously, there was some fallback behaviour which meant that requests
for a non-existent recruitment cycle year would silently fallback to
using the current cycle. However, this was implicit behaviour that
wasn't obvious.

Instead, I've changed the handling so that GET requests to
`/recruitment_cycles/current/providers` return providers in the current
recruitment cycle year, and requests for non-existent years return a 404
Not Found error.

I've updated the Swagger documentation accordingly.

Note: it looks like the Swagger JSON file was out of sync with the RSwag
specs. So when I ran `rake rswag:specs:swaggerize` a couple of other
changes were made to the file too.