Handle requests for courses in the current year
This commit tightens up the handling of requests for courses. The
changes made in this commit closely resemble those made to providers in
the previous commit.

Previously, there was some fallback behaviour which meant that requests
for a non-existent recruitment cycle year would silently fallback to
using the current cycle. However, this was implicit behaviour that
wasn't obvious.

Instead, I've changed the handling so that GET requests to
`/recruitment_cycles/current/courses` return providers in the current
recruitment cycle year, and requests for non-existent years return a 404
Not Found error.

I've updated the Swagger documentation accordingly.