[1423] Show schools even if there is only one
I think the reasoning behind this might be because previously if the provider only had one school that would always be called main site, which didn't mean a lot. This might still be the case for some providers who haven't changed it.

We have decided to remove this restriction as it’s confusing providers who are expecting to see their school listed on Find.