[1749] Dynamic back links when changing TDA to non TDA
When creating a TDA course we default certain attributes on the course. When a user changes a TDA course to a non TDA course, we step the user back through the questions which were previously defaulted.

We need to apply similar logic to the back links.

This commit is for changing a course which already exists in the databse.