Redirect away from adding more A level subjects when maximum A levels
If the user add 4 levels and somehow (manually) tries to enter on the
add A level page - we should redirect away.

If the user enters manually on the What A level is required or via back
in the browser and try to create another A level, the user will be
redirect to the list as there are already 4 A levels.