Add provider & accrediting_provider views course
The course show page for find and publish(preview page) has too much
content. This commit tries to reduce this content by putting the
provider of the course and the accrediting provider of the course into
their own views.

For this to happen we also needed to change the logic of the
redirects/back links when adding a provider through the course show
page. You can do this when you have a draft course but without a
provider.