[2016] Prevent the creation of a user which already exists under the provider
With out current implementation a user can add another user even if they already exist under that provider. This then updates the already existing user with the new first name and last name.

This will prevent submission of the "/new" page if the email address that is inputted already exists in the db for the provider.