Refactor goto methods for about org form redirect
There are a few `goto` methods that tell the about your organisation
form where to redirect the user after update or when the user clicks the
back links.

Basically, both back links and update success methods can go to the same
path based on the `goto` param.

Because of this and the ever growing of these `goto` params I moved some
logic that calculates the redirection in the
about_your_organisation_form. This way it can be tested properly.