Give a course being rolled over an enrichment
This is so that a manually rolled over course will have a content_status of rolled_over instead of draft.

When we manually rollover a course it has a content_status of draft but we want a manually rolled over course to have a content_status of rolled_over.

They are currently draft because they don't have any enrichments and in these cases we default the content_status to draft.

The code that causes the content_status to get set to rolled_over is never hit in content_service_status.rb because there is no enrichment to dup.