Redesign visa sponsorship on course page
The International students section is going to be removed from the
publis/find course show pages. This is part of a redesign of the course
page to make it easier for users to view the information without being
overwhelmed

The visa sponsorship content will now be in a drop-down details tag. The
content inside has also been tweaked.