Update results format for the course index page
This is part of a redesign of the course pages

It also includes a refactor of some methods that are used to display the
course information on the results page as well on the details page. This
way we can change the content in 1 place and have it mirrored in 2 places
and will help with content consistency.