[2124] Add tests for the funding column
Essentially duplicating the tests which are already testing the current implementation (funding_type) but testing with the feature flag activated.

Some files have been duplicated to avoid bloating the body of the file. Where this has been done, the filename has been updated to reflect the fact that the feature flag is inactive in the test and a comment has been added to the top of the file to remind us the whole file can be deleted when the feature flag and previous funding_type implementation is removed.

Some of these new files were failing on the linter, the original file had linting exclusion in place, so the same pattern was followed.