Show specific no results message when searching undergraduate courses
When Teacher Degree Apprenticeships launches it is likely there will
be no or very little courses available.

This may cause a dead end on the search results page,
which isn’t a good user journey.

Adding a more detailed explanation could help users.

This is a short term solution and there are talks for a long term
solution if this problem persists.