Unskip tests
These tests were previously failing in CI because Find was closed, we temporarily skipped them to get Find open.

Now that Find is open, we are unskipping them and adapting them for the future. There is no need to reference the 2025 cycle any more as these tests will apply to all future cycles.

All code references restricting TDA to the 2025 cycle should be removed too.