Implement the visa sponsorship filter in the new search results
I added

Course Search form which will be the form object responsible to handle
all params (old and new) and also typecast properly to the course
search service.

Course query logger that is responsible to show which filters the
logic is applying, print the query and also explain the query if we
need.