Move courses query logger to `CoursesQuery::Logger`
I created this class because of my own necessity.

To see which filters is applied, inspect better in the log the
whole query and also to explain to understand the query performance.

This is already helping me to see if the code applied the filter or not,
 see the query in the midst of the log chaos and also explain and
 understand a performance issue problem with the search (this one and
the one is live)

 There is a problem with the current search related to this
 associations:

 courses <-> course_subjects <-> subjects

Investigating ... the reason is because of the after_initialize block
in the course model (it ignores any Course.includes we do).